@@ -1,79 +1,72 @@
 {
-  "name": "apimah",
+  "name": "marketplace-api",
   "version": "1.0.0",
-  "description": "API para la web de mah",
-  "main": "src/index.js",
+  "description": "API para la tienda de TechMo",
+  "main": "src/app.js",
   "scripts": {
+    "start": "nodemon src/app.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "migrateDev": "node_modules/.bin/sequelize db:create --env test && node_modules/.bin/sequelize db:migrate --env test && node_modules/.bin/sequelize db:seed:all --env test && yarn run dev",
     "migratePG": "node_modules/.bin/sequelize db:migrate --env testpsql && node_modules/.bin/sequelize db:seed:all --env testpsql && yarn run dev",
-    "migrateProd": "node_modules/.bin/sequelize db:migrate --env production",
-    "dev": "nodemon src/index.js"
+    "migrateProd": "node_modules/.bin/sequelize db:migrate --env production"
   },
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@bitbucket.org/mfliri_ao/apicinesis.git"
+    "url": "https://github.com/er-mah/marketplace"
   },
   "keywords": [
-    "mah",
-    "mi",
-    "auto",
-    "hoy",
-    "mendoza",
-    "argentina",
+    "global",
+    "techmo",
+    "facil",
+    "rapidez",
     "tratamiento",
     "calidad"
   ],
-  "author": "As-One",
+  "author": "As-One, cuxs, estebanarivasv",
   "license": "ISC",
-  "homepage": "https://bitbucket.org/mfliri_ao/apicinesis#readme",
+  "homepage": "https://github.com/er-mah/marketplace",
   "dependencies": {
-    "@sendgrid/mail": "^6.2.1",
+    "@sendgrid/mail": "^7.7.0",
     "apollo-server-express": "^1.3.2",
     "bcrypt-nodejs": "^0.0.3",
-    "body-parser": "^1.18.2",
-    "casual": "^1.5.19",
-    "cors": "^2.8.4",
-    "dotenv": "^6.0.0",
-    "eslint": "^4.12.1",
-    "express": "^4.16.2",
-    "express-graphql": "^0.6.12",
-    "express-jwt": "^5.3.0",
-    "graphql": "^0.12.3",
+    "body-parser": "^1.20.1",
+    "casual": "^1.6.2",
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.2.0",
+    "express": "^4.18.2",
+    "express-graphql": "^0.12.0",
+    "express-jwt": "^8.4.1",
+    "graphql": "^16.6.0",
     "graphql-errors": "^2.1.0",
-    "graphql-relay": "^0.5.4",
-    "graphql-sequelize": "^5.6.1",
-    "graphql-subscriptions": "^0.5.6",
-    "jsonwebtoken": "^8.1.0",
-    "jwt-decode": "^2.2.0",
-    "lodash": "^4.17.4",
-    "mercadolibre": "^0.0.13",
-    "method-override": "^2.3.10",
-    "moment": "^2.19.3",
-    "morgan": "^1.9.1",
+    "graphql-relay": "^0.10.0",
+    "graphql-sequelize": "^9.5.1",
+    "graphql-subscriptions": "^2.0.0",
+    "jsonwebtoken": "^9.0.0",
+    "jwt-decode": "^3.1.2",
+    "lodash": "^4.17.21",
+    "method-override": "^3.0.0",
+    "moment": "^2.29.4",
+    "morgan": "^1.10.0",
     "multer": "^1.3.0",
-    "mysql2": "^1.5.1",
-    "newrelic": "^4.1.3",
-    "node-cache": "^4.2.0",
-    "node-fetch": "^2.2.0",
-    "node-schedule": "^1.3.0",
-    "node-thumbnail": "^0.14.0",
-    "nodemon": "^1.12.1",
-    "pg": "^7.4.1",
-    "python-shell": "^0.5.0",
-    "sequelize": "^4.38.0",
-    "sequelize-auto": "^0.4.29",
-    "sequelize-cli": "^4.0.0",
-    "sharp": "^0.20.0",
-    "simple-oauth2": "^2.2.1",
-    "slugify": "^1.3.4",
-    "socket.io": "^2.0.4",
-    "socketio-jwt": "^4.5.0",
+    "mysql2": "^3.1.2",
+    "newrelic": "^9.10.1",
+    "node-cache": "^5.1.2",
+    "node-fetch": "^3.3.0",
+    "node-schedule": "^2.1.1",
+    "node-thumbnail": "^0.15.0",
+    "nodemon": "^2.0.20",
+    "pg": "^8.9.0",
+    "python-shell": "^5.0.0",
+    "sequelize": "^6.28.0",
+    "sequelize-auto": "^0.8.8",
+    "sequelize-cli": "^6.6.0",
+    "sharp": "^0.31.3",
+    "simple-oauth2": "^5.0.0",
+    "slugify": "^1.6.5",
+    "socket.io": "^4.6.0",
+    "socketio-jwt": "^4.6.2",
     "split-object": "^2.1.1",
     "subscriptions-transport-ws": "^0.9.4"
-  },
-  "devDependencies": {
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-plugin-import": "^2.8.0"
   }
 }
